requirements.txt
azure-functions==1.17.0
azure-storage-blob==12.19.0
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
openpyxl==3.1.2
python-multipart==0.0.6
